fix:Dealcontroller } 수정(S10P12A302-147)

+++ backend/src/main/java/com/ssafy/oywo/dto/DealRequestsDto.java
@@ -15,7 +15,7 @@ public class DealRequestsDto {
     private String content;
     private Long reward;
     private RewardType rewardType;
-    private Long complaint;
+    private int complaint;
     private Status status;
     private String dealType;
     private LocalDateTime expireAt;
