+++ backend/src/main/java/com/ssafy/oywo/dto/DealResponseDto.java
@@ -19,7 +19,7 @@ public class DealResponseDto {
     private String content;
     private Long reward;
     private RewardType rewardType;
-    private Long complaint;
+    private int complaint;
     private Status status;
     private String dealType;
     private LocalDateTime expireAt;
